2021 12 14 refactor deployment

+++ shell.nix
@@ -2,7 +2,7 @@ let
  pkgs = import <nixpkgs> {};
 
    flush-all = pkgs.writeShellScriptBin "flush-all" ''
-    rm -rf node_modules
+    # rm -rf node_modules
     rm -rf artifacts
     rm -rf cache
     rm -rf typechain
@@ -10,9 +10,26 @@ let
     yarn install
   '';
 
-  cut-dist = pkgs.writeShellScriptBin "cut-dist" ''
-    flush-all
+   deploy-rain = pkgs.writeShellScriptBin "deploy-rain" ''
+    npx hardhat run scripts/deploy-rain.ts --network ''$1
+  '';
 
+   create-trust = pkgs.writeShellScriptBin "create-trust" ''
+    export TrustFactory=''$1
+    npx hardhat run scripts/create-trust.ts --network ''$2
+  '';
+
+   deploy-verify = pkgs.writeShellScriptBin "deploy-verify" ''
+    export AdminAddress=''$1
+    npx hardhat run scripts/deploy-verify-tier.ts --network ''$2
+  '';
+
+  cut-dist = pkgs.writeShellScriptBin "cut-dist" ''
+    if [ -z "$1" ]; then echo "ERROR: Missing commit argument. Exiting..."; exit 0; fi
+    flush-all
+    commit=$1; copy-commit $1
+    sed -i '$s/.*/COMMIT='$commit'/' .env
+    sed -i '$s/.*/COMMIT='$commit'/' .env.example
     hardhat compile --force
     rm -rf dist
     mkdir -p "dist"
@@ -23,6 +40,11 @@ let
     cp -r "solt" "dist"
   '';
 
+  copy-commit = pkgs.writeShellScriptBin "copy-commit" ''
+    (cd rain-protocol; git checkout develop; git pull; git checkout $1)
+    rm -rf contracts; cp -r rain-protocol/contracts ./
+  '';
+
   solt-the-earth = pkgs.writeShellScriptBin "solt-the-earth" ''
     mkdir -p solt
     find contracts -type f -not -path 'contracts/test/*' | xargs -i solt write '{}' --npm --runs 100000
@@ -31,30 +53,19 @@ let
     mv solc-* solt
   '';
   
-  deploy-rain = pkgs.writeShellScriptBin "deploy-rain" ''
-    yarn deploy-rain --network
-  '';
-
-  deploy-rain-mumbai = pkgs.writeShellScriptBin "deploy-rain-mumbai" ''
-    yarn deploy-rain --network mumbai
-  '';
-
-  deploy-rain-reef-testnet = pkgs.writeShellScriptBin "deploy-rain-reef-testnet" ''
-    yarn deploy-rain-reef --network reef_testnet
-  '';
-  
 in
 pkgs.stdenv.mkDerivation {
  name = "shell";
  buildInputs = [
   pkgs.nodejs-14_x
   pkgs.jq
-  deploy-rain-mumbai
-  deploy-rain-reef-testnet
   deploy-rain
+  create-trust
+  deploy-verify
   flush-all
   cut-dist
   solt-the-earth
+  copy-commit
  ];
 
  shellHook = ''
